feat(frontend): refactor useTodos hook with responsibility separation
Major refactoring to improve maintainability and testability:

**New Architecture:**
- `todo-optimistic-utils.ts`: Pure utility functions for optimistic updates
- `useTodosState.ts`: State management and setters (59 lines)
- `useTodoMutations.ts`: CRUD operations with optimistic updates (168 lines)
- `useTodosFilter.ts`: Filtering and count calculations (49 lines)
- `useTodos.ts`: Main hook integrating all functionality (89 lines)

**Benefits:**
- Single Responsibility Principle: Each hook has one clear purpose
- Better testability: Individual concerns can be tested in isolation
- Improved reusability: Components can import specific functionality
- Reduced complexity: Down from 250 lines to organized, focused modules
- Enhanced maintainability: Easier to understand and modify

**Original Implementation:**
- Backed up to `useTodos.original.ts` for reference
- All existing functionality preserved with same public API
